fix: resolve remaining logger and type errors
- Fixed pkg/monitoring/tracing.go: Wrapped all logger fields with zap functions
- Fixed pkg/monitoring/metrics.go, prometheus.go: Removed unused key variables
- Fixed pkg/service/registry.go: Fixed logger calls with proper zap wrappers
- Fixed pkg/service/web3.go: Added GetEthClient() method to ChainClient
- Fixed pkg/web3/chain.go: Added getter method for ethclient
- Fixed pkg/optimization/resource_optimizer.go: Fixed float to uint64 conversion
- Fixed pkg/optimization/service.go: Fixed time.Duration type and added import
- Fixed pkg/optimization/caching.go: Removed unused fmt import
- Fixed pkg/core/microkernel.go: Wrapped all logger fields with zap functions

Progress: ~95% of compilation errors fixed

+++ pkg/optimization/resource_optimizer.go
@@ -301,7 +301,7 @@ func (ro *ResourceOptimizer) GetOptimizationRecommendations() []string {
 		firstMetric := ro.memoryTrends[0]
 		lastMetric := ro.memoryTrends[len(ro.memoryTrends)-1]
 
-		if lastMetric.HeapAlloc > firstMetric.HeapAlloc*1.5 {
+		if lastMetric.HeapAlloc > uint64(float64(firstMetric.HeapAlloc)*1.5) {
 			recommendations = append(recommendations, "Potential memory leak detected - heap allocation increased significantly")
 		}
 	}
